Rename: applications/ -> application/
