+++ application/repositories/exerciseRepository.js
