+++ application/useCases/findExerciseByUser.js
